Compiled AddPatient, PatientDirectory and PatientProfile into separate scripts in screens folder
