+++ package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@react-navigation/native": "^7.0.14",
+    "@react-navigation/native-stack": "^7.2.0",
     "@react-navigation/stack": "^7.1.0",
     "expo": "^52.0.0",
     "expo-linear-gradient": "^14.0.1",
