+++ types.ts
@@ -1,6 +1,24 @@
-export type patient = {
-    id: number,
+import { NativeStackNavigationProp } from "@react-navigation/native-stack";
+import { RouteProp } from "@react-navigation/native";
+
+export type Patient = {
+    id: string,
     firstName: string,
     lastName: string,
     level: number,
 }
+
+// List of screens that are allowed
+export type RootStackParamList = {
+    PatientDirectory: undefined;
+    AddPatient: undefined;
+    PatientProfile: {
+        patient: Patient | undefined
+    };
+}
+
+
+// Used to auto fill navigation tools with allowed scenes
+export type ScreenNavigationProp = NativeStackNavigationProp<RootStackParamList>;
+
+export type EditScreenRouteProp = RouteProp<RootStackParamList, 'PatientProfile'>;
